@@ -1,13 +1,13 @@
 package com.vueespring;
 
-import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 
 import java.util.TimeZone;
 
-@MapperScan("com.vueespring.mapper")
 @SpringBootApplication
+@EnableRedisRepositories
 public class VueeSpringApplication {
 
     public static void main(String[] args) {
