@@ -1,24 +1,21 @@
 package com.vueespring.controller;
 
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.vueespring.entity.WebEntity.UserVoeEntity;
-import com.vueespring.mapper.UserVoeTableMapper;
-import com.vueespring.service.IUserVoeTableService;
-import com.vueespring.service.QuartzService;
-import com.vueespring.shiro.JwtUtils;
+import cn.dev33.satoken.stp.StpUtil;
+import cn.dev33.satoken.util.SaResult;
+import com.vueespring.entity.WebEntity.UserEntity;
+import com.vueespring.service.UserService;
 import com.vueespring.utils.JsonResult;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.web.bind.annotation.*;
-import cn.hutool.crypto.SecureUtil;
+
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.time.LocalDateTime;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
 /**
  * <p>
@@ -30,59 +27,81 @@ import java.util.Map;
  */
 @RestController
 @Slf4j
+@CrossOrigin
 public class AccountController {
     @Autowired
-    IUserVoeTableService iUserVoeTableService;
+    UserService userService;
     @Autowired
-    JwtUtils jwtUtils;
-    @Autowired
-    UserVoeTableMapper userVoeTableMapper;
-    @Autowired
-    QuartzService quartzService;
+    MongoTemplate mongoTemplate;
     @RequestMapping("/login")
-    public JsonResult login(@RequestParam(value="username")String username,
-                            @RequestParam(value="password")String password,
-                            HttpServletResponse response) {
-        QueryWrapper<UserVoeEntity> wrapper = new QueryWrapper<UserVoeEntity>().eq("username",username);
-        log.debug(username);
-        log.debug(password);
-        UserVoeEntity voeEntity = iUserVoeTableService.getOne(wrapper);
-        log.debug(voeEntity.toString());
-        if(!voeEntity.getPassword().equals(SecureUtil.md5(password))){
-            return new JsonResult().error("密码错误");
+    public SaResult login(@RequestParam(value="username")String username,
+                          @RequestParam(value="password")String password,
+                          @RequestParam("remember") Boolean remember,
+    HttpServletRequest request) {
+        UserEntity user = userService.getuserByUsernamePassword(username,password);
+        if(user==null){
+            if(userService.getUserByUsername(username)!=null){
+                return SaResult.error("密码错误");
+            }
+            else return SaResult.error("用户名错误");
         }
-        else {
-            String token = jwtUtils.generateToken(voeEntity.getId());
-            LocalDateTime time = LocalDateTime.now().plusSeconds(60*60*48);
-            response.setCharacterEncoding("utf-8");
-            response.setHeader("Token",token);
-            response.setHeader("Access-Control-Expose-Headers", "token");
-            Map<String,Object> map = new HashMap<String, Object>();
-            map.put("username",voeEntity.getUsername());
-            map.put("message","Login Successfully");
-            map.put("role",voeEntity.getPermission());
-            map.put("nextexpiretime",time);
-            return new JsonResult().ok(map);
+        String id = user.getId();
+        if(remember){
+            StpUtil.login(id,true);
+        }else {
+            StpUtil.login(id);
         }
+        return new SaResult()
+                .setCode(200)
+                .setData(user)
+                .setMsg("登录成功");
     }
     @GetMapping("/register")
-    public JsonResult register(@RequestParam String username,
-                               @RequestParam String password,
-                               HttpServletResponse response){
-        QueryWrapper<UserVoeEntity> wrapper = new QueryWrapper<UserVoeEntity>().eq("username",username);
-        if(iUserVoeTableService.count(wrapper)>0){
-            return new JsonResult().error("已有账号，可登陆");
-        }else {
-            if(iUserVoeTableService.save(new UserVoeEntity(null,username,SecureUtil.md5(password),"user"))){
-                return new JsonResult().ok("注册成功");
+    public SaResult register(@RequestParam String username,
+                             @RequestParam String password,
+                             HttpServletRequest request){
+        UserEntity user = userService.getUserByUsername(username);
+        UserEntity userEntity = new UserEntity();
+        userEntity.setUsername(username);
+        userEntity.setPassword(password);
+        userEntity.setPermission("user");
+        if(user==null){
+            UserEntity insert = mongoTemplate.insert(userEntity);
+            if(insert!=null){
+                return SaResult.ok("注册成功");
             }else {
-                return new JsonResult().error("注册失败");
+                return SaResult.error("注册失败");
             }
         }
+        else {
+            return SaResult.error("用户名已存在");
+        }
     }
+
     @GetMapping("/logout")
-    public JsonResult logout(){
-        SecurityUtils.getSubject().logout();
-        return new JsonResult().ok("退出成功");
+    public SaResult logout(
+            HttpServletRequest request){
+        StpUtil.logout();
+        return SaResult.ok("已登出");
+    }
+    @GetMapping("/setAlertToken")
+    public SaResult setToken(
+            @RequestParam("alertToken") String token,
+            HttpServletRequest request) {
+        String loginId = StpUtil.getLoginIdAsString();
+        UserEntity user = userService.getUserById(loginId);
+        Query query = new Query(Criteria.where("id")
+                .is(user.getId()));
+        Update update = new Update();
+        update.set("alertToken",token);
+
+        if(user==null){
+            return SaResult.error("未找到用户，token无法修改");
+        }
+
+        user.setAlertToken(token);
+
+        mongoTemplate.updateFirst(query,update,UserEntity.class);
+        return SaResult.ok("update success");
     }
 }
