@@ -1,22 +1,22 @@
 package com.vueespring.service;
 
-import com.vueespring.entity.Thingstable;
-import com.vueespring.entity.WebEntity.Item.ItemVOEntity;
-import com.vueespring.entity.WebEntity.UserVoeEntity;
+import com.vueespring.entity.ThingEnity;
+import com.vueespring.entity.WebEntity.Item.Itemtity;
+import com.vueespring.entity.WebEntity.UserEntity;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
 
 
 public interface ThingService {
-    void creatitem(Thingstable thing, Scheduler scheduler) throws SchedulerException;
+    void creatitem(ThingEnity thing, Scheduler scheduler) throws SchedulerException;
 
-    void startitem(Thingstable thingstable, Scheduler scheduler) throws SchedulerException;
+    void startitem(ThingEnity thingEnity, Scheduler scheduler) throws SchedulerException;
 
-    void pausething(Thingstable thing, Scheduler scheduler);
+    void pausething(ThingEnity thing, Scheduler scheduler);
 
-    void delthing(Thingstable thing, Scheduler scheduler) throws SchedulerException;
+    void delthing(ThingEnity thing, Scheduler scheduler) throws SchedulerException;
 
-    String checkAndSetStatus(Thingstable thing);
+    String checkAndSetStatus(ThingEnity thing);
 
-    Thingstable getThingByVoe(ItemVOEntity itemVOEntity, String userid, UserVoeEntity userinfo);
+    ThingEnity getThingByVoe(Itemtity itemtity, String userid, UserEntity userinfo);
 }
