@@ -5,17 +5,22 @@ import cn.hutool.http.HttpUtil;
 import com.vueespring.service.FWAlertJobService;
 import lombok.extern.slf4j.Slf4j;
 import org.quartz.JobDataMap;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
+import javax.annotation.Resource;
 import java.util.Map;
 
 @Slf4j
 @Service
 public class FWAlertJobServiceImpl implements FWAlertJobService {
+    @Resource
+    RestTemplate restTemplate;
     @Override
-    public void sendFWAlert(Map<String, Object> map, String token) {
+    public void sendFWAlert(Map<String, Object> map) {
         log.debug(String.valueOf(map));
-        HttpResponse res = HttpUtil.createGet("https://fwalert.com/"+ token)
+        HttpResponse res = HttpUtil.createGet("https://fwalert.com/"+ map.get("alertToken"))
                         .form(map)
                         .execute();
         log.debug(String.valueOf(res));
@@ -28,5 +33,6 @@ public class FWAlertJobServiceImpl implements FWAlertJobService {
         map.put("type", jobDataMap.get("type"));
         map.put("tag", jobDataMap.get("tag"));
         map.put("name", jobDataMap.get("name"));
+        map.put("alertToken",jobDataMap.get("alertToken"));
     }
 }
